refact: render transfer details form in PayExpenseModal
